refactor(ui): split analytics.tsx into chart components and audit table #164
Extract RingChart, BarChart, HorizontalBarChart, and AuditLogTable into
src/app/components/analytics/ to bring analytics.tsx from 666 lines down
to 269 lines (under the 400-line warning threshold). Pure refactor with
no logic or UI changes.
